requirements.txt
asn1crypto==1.2.0
azure-common==1.1.23
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto3==1.9.253
botocore==1.12.253
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
docutils==0.15.2
future==0.18.1
idna==2.8
ijson==2.5.1
jmespath==0.9.4
oscrypto==1.1.0
psycopg2==2.8.4
pycparser==2.19
pycryptodomex==3.9.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
snowflake-connector-python==2.0.2
urllib3==1.25.6
